feat(v8): Add @sentry/aws-serverless package
