+++ .craft.yml
@@ -89,8 +89,8 @@ targets:
 
   ## 5. Node-based Packages
   - name: npm
-    id: '@sentry/serverless'
-    includeNames: /^sentry-serverless-\d.*\.tgz$/
+    id: '@sentry/aws-serverless'
+    includeNames: /^sentry-aws-serverless-\d.*\.tgz$/
   - name: npm
     id: '@sentry/google-cloud'
     includeNames: /^sentry-google-cloud-\d.*\.tgz$/
