+++ .github/ISSUE_TEMPLATE/bug.yml
@@ -34,6 +34,7 @@ body:
         - '@sentry/astro'
         - '@sentry/angular'
         - '@sentry/angular-ivy'
+        - '@sentry/aws-serverless'
         - '@sentry/bun'
         - '@sentry/deno'
         - '@sentry/ember'
@@ -42,7 +43,6 @@ body:
         - '@sentry/node'
         - '@sentry/react'
         - '@sentry/remix'
-        - '@sentry/serverless'
         - '@sentry/google-cloud'
         - '@sentry/svelte'
         - '@sentry/sveltekit'
