+++ dev-packages/e2e-tests/verdaccio-config/config.yaml
@@ -134,7 +134,7 @@ packages:
     unpublish: $all
     # proxy: npmjs # Don't proxy for E2E tests!
 
-  '@sentry/serverless':
+  '@sentry/aws-serverless':
     access: $all
     publish: $all
     unpublish: $all
