+++ packages/aws-serverless/LICENSE
