+++ packages/aws-serverless/src/debug-build.ts
