+++ packages/aws-serverless/src/index.awslambda.ts
