+++ packages/aws-serverless/test/__mocks__/dns.ts
