+++ packages/aws-serverless/test/awsservices.test.ts
@@ -1,4 +1,4 @@
-import { NodeClient, createTransport, setCurrentClient } from '@sentry/node-experimental';
+import { NodeClient, createTransport, setCurrentClient } from '@sentry/node';
 import * as AWS from 'aws-sdk';
 import * as nock from 'nock';
 
@@ -8,9 +8,9 @@ import { awsServicesIntegration } from '../src/awsservices';
 const mockSpanEnd = jest.fn();
 const mockStartInactiveSpan = jest.fn(spanArgs => ({ ...spanArgs }));
 
-jest.mock('@sentry/node-experimental', () => {
+jest.mock('@sentry/node', () => {
   return {
-    ...jest.requireActual('@sentry/node-experimental'),
+    ...jest.requireActual('@sentry/node'),
     startInactiveSpan: (ctx: unknown) => {
       mockStartInactiveSpan(ctx);
       return { end: mockSpanEnd };
