--- packages/serverless/test/gcpfunction.test.ts
@@ -1,595 +1,0 @@
-import * as domain from 'domain';
-
-import type { Event, Integration } from '@sentry/types';
-
-import { SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN, SEMANTIC_ATTRIBUTE_SENTRY_SOURCE } from '@sentry/core';
-
-import { wrapCloudEventFunction, wrapEventFunction, wrapHttpFunction } from '../src/gcpfunction';
-import type {
-  CloudEventFunction,
-  CloudEventFunctionWithCallback,
-  EventFunction,
-  EventFunctionWithCallback,
-  HttpFunction,
-  Request,
-  Response,
-} from '../src/gcpfunction/general';
-
-import { init } from '../src/gcpfunction';
-
-const mockStartInactiveSpan = jest.fn((...spanArgs) => ({ ...spanArgs }));
-const mockStartSpanManual = jest.fn((...spanArgs) => ({ ...spanArgs }));
-const mockFlush = jest.fn((...args) => Promise.resolve(args));
-const mockWithScope = jest.fn();
-const mockCaptureMessage = jest.fn();
-const mockCaptureException = jest.fn();
-const mockInit = jest.fn();
-
-const mockScope = {
-  setTag: jest.fn(),
-  setContext: jest.fn(),
-  addEventProcessor: jest.fn(),
-  setSDKProcessingMetadata: jest.fn(),
-};
-
-const mockSpan = {
-  end: jest.fn(),
-};
-
-jest.mock('@sentry/node-experimental', () => {
-  const original = jest.requireActual('@sentry/node-experimental');
-  return {
-    ...original,
-    init: (options: unknown) => {
-      mockInit(options);
-    },
-    startInactiveSpan: (...args: unknown[]) => {
-      mockStartInactiveSpan(...args);
-      return mockSpan;
-    },
-    startSpanManual: (...args: unknown[]) => {
-      mockStartSpanManual(...args);
-      mockSpan.end();
-      return original.startSpanManual(...args);
-    },
-    getCurrentScope: () => {
-      return mockScope;
-    },
-    flush: (...args: unknown[]) => {
-      return mockFlush(...args);
-    },
-    withScope: (fn: (scope: unknown) => void) => {
-      mockWithScope(fn);
-      fn(mockScope);
-    },
-    captureMessage: (...args: unknown[]) => {
-      mockCaptureMessage(...args);
-    },
-    captureException: (...args: unknown[]) => {
-      mockCaptureException(...args);
-    },
-  };
-});
-
-describe('GCPFunction', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  async function handleHttp(fn: HttpFunction, trace_headers: { [key: string]: string } | null = null): Promise<void> {
-    let headers: { [key: string]: string } = { host: 'hostname', 'content-type': 'application/json' };
-    if (trace_headers) {
-      headers = { ...headers, ...trace_headers };
-    }
-    return new Promise((resolve, _reject) => {
-      const d = domain.create();
-      const req = {
-        method: 'POST',
-        url: '/path?q=query',
-        headers: headers,
-        body: { foo: 'bar' },
-      } as Request;
-      const res = { end: resolve } as Response;
-      d.on('error', () => res.end());
-      d.run(() => process.nextTick(fn, req, res));
-    });
-  }
-
-  function handleEvent(fn: EventFunctionWithCallback): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const d = domain.create();
-      // d.on('error', () => res.end());
-      const context = {
-        eventType: 'event.type',
-        resource: 'some.resource',
-      };
-      d.on('error', reject);
-      d.run(() =>
-        process.nextTick(fn, {}, context, (err: any, result: any) => {
-          if (err != null || err != undefined) {
-            reject(err);
-          } else {
-            resolve(result);
-          }
-        }),
-      );
-    });
-  }
-
-  function handleCloudEvent(fn: CloudEventFunctionWithCallback): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const d = domain.create();
-      // d.on('error', () => res.end());
-      const context = {
-        type: 'event.type',
-      };
-      d.on('error', reject);
-      d.run(() =>
-        process.nextTick(fn, context, (err: any, result: any) => {
-          if (err != null || err != undefined) {
-            reject(err);
-          } else {
-            resolve(result);
-          }
-        }),
-      );
-    });
-  }
-
-  describe('wrapHttpFunction() options', () => {
-    test('flushTimeout', async () => {
-      const handler: HttpFunction = (_, res) => {
-        res.end();
-      };
-      const wrappedHandler = wrapHttpFunction(handler, { flushTimeout: 1337 });
-
-      await handleHttp(wrappedHandler);
-      expect(mockFlush).toBeCalledWith(1337);
-    });
-  });
-
-  describe('wrapHttpFunction()', () => {
-    test('successful execution', async () => {
-      const handler: HttpFunction = (_req, res) => {
-        res.statusCode = 200;
-        res.end();
-      };
-      const wrappedHandler = wrapHttpFunction(handler);
-      await handleHttp(wrappedHandler);
-
-      const fakeTransactionContext = {
-        name: 'POST /path',
-        op: 'function.gcp.http',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'route',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_http',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalledWith(2000);
-    });
-
-    test('capture error', async () => {
-      const error = new Error('wat');
-      const handler: HttpFunction = (_req, _res) => {
-        throw error;
-      };
-      const wrappedHandler = wrapHttpFunction(handler);
-
-      await handleHttp(wrappedHandler);
-
-      const fakeTransactionContext = {
-        name: 'POST /path',
-        op: 'function.gcp.http',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'route',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_http',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalled();
-    });
-
-    test('should not throw when flush rejects', async () => {
-      const handler: HttpFunction = async (_req, res) => {
-        res.statusCode = 200;
-        res.end();
-      };
-
-      const wrappedHandler = wrapHttpFunction(handler);
-
-      const request = {
-        method: 'POST',
-        url: '/path?q=query',
-        headers: { host: 'hostname', 'content-type': 'application/json' },
-        body: { foo: 'bar' },
-      } as Request;
-
-      const mockEnd = jest.fn();
-      const response = { end: mockEnd } as unknown as Response;
-
-      mockFlush.mockImplementationOnce(async () => {
-        throw new Error();
-      });
-
-      await expect(wrappedHandler(request, response)).resolves.toBeUndefined();
-      expect(mockEnd).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  // This tests that the necessary pieces are in place for request data to get added to event - the `RequestData`
-  // integration is included in the defaults and the necessary data is stored in `sdkProcessingMetadata`. The
-  // integration's tests cover testing that it uses that data correctly.
-  test('wrapHttpFunction request data prereqs', async () => {
-    init({});
-
-    const handler: HttpFunction = (_req, res) => {
-      res.end();
-    };
-    const wrappedHandler = wrapHttpFunction(handler);
-
-    await handleHttp(wrappedHandler);
-
-    const initOptions = (mockInit as unknown as jest.SpyInstance).mock.calls[0];
-    const defaultIntegrations = initOptions[0].defaultIntegrations.map((i: Integration) => i.name);
-
-    expect(defaultIntegrations).toContain('RequestData');
-
-    expect(mockScope.setSDKProcessingMetadata).toHaveBeenCalledWith({
-      request: {
-        method: 'POST',
-        url: '/path?q=query',
-        headers: { host: 'hostname', 'content-type': 'application/json' },
-        body: { foo: 'bar' },
-      },
-    });
-  });
-
-  describe('wrapEventFunction() without callback', () => {
-    test('successful execution', async () => {
-      const func: EventFunction = (_data, _context) => {
-        return 42;
-      };
-      const wrappedHandler = wrapEventFunction(func);
-      await expect(handleEvent(wrappedHandler)).resolves.toBe(42);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalledWith(2000);
-    });
-
-    test('capture error', async () => {
-      const error = new Error('wat');
-      const handler: EventFunction = (_data, _context) => {
-        throw error;
-      };
-      const wrappedHandler = wrapEventFunction(handler);
-      await expect(handleEvent(wrappedHandler)).rejects.toThrowError(error);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalled();
-    });
-  });
-
-  describe('wrapEventFunction() as Promise', () => {
-    test('successful execution', async () => {
-      const func: EventFunction = (_data, _context) =>
-        new Promise(resolve => {
-          setTimeout(() => {
-            resolve(42);
-          }, 10);
-        });
-      const wrappedHandler = wrapEventFunction(func);
-      await expect(handleEvent(wrappedHandler)).resolves.toBe(42);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalledWith(2000);
-    });
-
-    test('capture error', async () => {
-      const error = new Error('wat');
-      const handler: EventFunction = (_data, _context) =>
-        new Promise((_, reject) => {
-          setTimeout(() => {
-            reject(error);
-          }, 10);
-        });
-
-      const wrappedHandler = wrapEventFunction(handler);
-      await expect(handleEvent(wrappedHandler)).rejects.toThrowError(error);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalled();
-    });
-  });
-
-  describe('wrapEventFunction() with callback', () => {
-    test('successful execution', async () => {
-      const func: EventFunctionWithCallback = (_data, _context, cb) => {
-        cb(null, 42);
-      };
-      const wrappedHandler = wrapEventFunction(func);
-      await expect(handleEvent(wrappedHandler)).resolves.toBe(42);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalledWith(2000);
-    });
-
-    test('capture error', async () => {
-      const error = new Error('wat');
-      const handler: EventFunctionWithCallback = (_data, _context, cb) => {
-        cb(error);
-      };
-      const wrappedHandler = wrapEventFunction(handler);
-      await expect(handleEvent(wrappedHandler)).rejects.toThrowError(error);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalled();
-    });
-
-    test('capture exception', async () => {
-      const error = new Error('wat');
-      const handler: EventFunctionWithCallback = (_data, _context, _cb) => {
-        throw error;
-      };
-      const wrappedHandler = wrapEventFunction(handler);
-      await expect(handleEvent(wrappedHandler)).rejects.toThrowError(error);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-    });
-  });
-
-  test('marks the captured error as unhandled', async () => {
-    const error = new Error('wat');
-    const handler: EventFunctionWithCallback = (_data, _context, _cb) => {
-      throw error;
-    };
-    const wrappedHandler = wrapEventFunction(handler);
-    await expect(handleEvent(wrappedHandler)).rejects.toThrowError(error);
-
-    expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-
-    const scopeFunction = mockCaptureException.mock.calls[0][1];
-    const event: Event = { exception: { values: [{}] } };
-    let evtProcessor: ((e: Event) => Event) | undefined = undefined;
-    scopeFunction({ addEventProcessor: jest.fn().mockImplementation(proc => (evtProcessor = proc)) });
-
-    expect(evtProcessor).toBeInstanceOf(Function);
-    // @ts-expect-error just mocking around...
-    expect(evtProcessor(event).exception.values[0].mechanism).toEqual({
-      handled: false,
-      type: 'generic',
-    });
-  });
-
-  test('wrapEventFunction scope data', async () => {
-    const handler: EventFunction = (_data, _context) => 42;
-    const wrappedHandler = wrapEventFunction(handler);
-    await handleEvent(wrappedHandler);
-    expect(mockScope.setContext).toBeCalledWith('gcp.function.context', {
-      eventType: 'event.type',
-      resource: 'some.resource',
-    });
-  });
-
-  describe('wrapCloudEventFunction() without callback', () => {
-    test('successful execution', async () => {
-      const func: CloudEventFunction = _context => {
-        return 42;
-      };
-      const wrappedHandler = wrapCloudEventFunction(func);
-      await expect(handleCloudEvent(wrappedHandler)).resolves.toBe(42);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.cloud_event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_cloud_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalledWith(2000);
-    });
-
-    test('capture error', async () => {
-      const error = new Error('wat');
-      const handler: CloudEventFunction = _context => {
-        throw error;
-      };
-      const wrappedHandler = wrapCloudEventFunction(handler);
-      await expect(handleCloudEvent(wrappedHandler)).rejects.toThrowError(error);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.cloud_event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_cloud_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalled();
-    });
-  });
-
-  describe('wrapCloudEventFunction() with callback', () => {
-    test('successful execution', async () => {
-      const func: CloudEventFunctionWithCallback = (_context, cb) => {
-        cb(null, 42);
-      };
-      const wrappedHandler = wrapCloudEventFunction(func);
-      await expect(handleCloudEvent(wrappedHandler)).resolves.toBe(42);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.cloud_event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_cloud_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalledWith(2000);
-    });
-
-    test('capture error', async () => {
-      const error = new Error('wat');
-      const handler: CloudEventFunctionWithCallback = (_context, cb) => {
-        cb(error);
-      };
-      const wrappedHandler = wrapCloudEventFunction(handler);
-      await expect(handleCloudEvent(wrappedHandler)).rejects.toThrowError(error);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.cloud_event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_cloud_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-      expect(mockSpan.end).toBeCalled();
-      expect(mockFlush).toBeCalled();
-    });
-
-    test('capture exception', async () => {
-      const error = new Error('wat');
-      const handler: CloudEventFunctionWithCallback = (_context, _cb) => {
-        throw error;
-      };
-      const wrappedHandler = wrapCloudEventFunction(handler);
-      await expect(handleCloudEvent(wrappedHandler)).rejects.toThrowError(error);
-
-      const fakeTransactionContext = {
-        name: 'event.type',
-        op: 'function.gcp.cloud_event',
-        attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: 'component',
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.function.serverless.gcp_cloud_event',
-        },
-      };
-
-      expect(mockStartSpanManual).toBeCalledWith(fakeTransactionContext, expect.any(Function));
-      expect(mockCaptureException).toBeCalledWith(error, expect.any(Function));
-    });
-  });
-
-  test('wrapCloudEventFunction scope data', async () => {
-    const handler: CloudEventFunction = _context => 42;
-    const wrappedHandler = wrapCloudEventFunction(handler);
-    await handleCloudEvent(wrappedHandler);
-    expect(mockScope.setContext).toBeCalledWith('gcp.function.context', { type: 'event.type' });
-  });
-
-  describe('init()', () => {
-    test('calls Sentry.init with correct sdk info metadata', () => {
-      init({});
-
-      expect(mockInit).toBeCalledWith(
-        expect.objectContaining({
-          _metadata: {
-            sdk: {
-              name: 'sentry.javascript.serverless',
-              integrations: ['GCPFunction'],
-              packages: [
-                {
-                  name: 'npm:@sentry/serverless',
-                  version: expect.any(String),
-                },
-              ],
-              version: expect.any(String),
-            },
-          },
-        }),
-      );
-    });
-  });
-});
